Split NotificationController into multiple classes for maintainability
